% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_functions.R
\name{alpha_metropolis}
\alias{alpha_metropolis}
\title{Metropolis-Hastings update of distortion probability for \eqn{\ell}th field.}
\usage{
alpha_metropolis(current, linkage, field, thetas_l, a, b, proposal, drift)
}
\arguments{
\item{current}{A float. The current value of \eqn{\alpha_\ell} in the Markov
chain.}

\item{linkage}{An integer vector. The cluster labels for each record.}

\item{field}{A character vector. The values for the \eqn{\ell^{th}} field of
the records.}

\item{thetas_l}{A matrix. The first column of \code{thetas_l} must contain
all the possible values that the \eqn{\ell^{th}} field can
take (i.e. the unique values of \code{field}), and the second
column must contain the corresponding frequencies of each value
in the whole set of records.}

\item{a}{A float. The \eqn{\alpha} shape parameter of the beta prior
for \eqn{\alpha_\ell}.}

\item{b}{A float. The \eqn{\beta} shape parameter of the beta prior
for \eqn{\alpha_\ell}.}

\item{proposal}{A string. Specifies the proposal distribution to be used in
generating a new value for \eqn{\alpha_\ell}. Currently only
accepts "RRW" for reflected random walk.}

\item{drift}{A float. A control parameter for tuning the variance of the
proposal distribution. larger values will lead to less
correlated proposals.}
}
\value{
The next value of \eqn{\alpha_\ell} in the Markov chain.
}
\description{
Implementation of the Metropolis-Hastings update step for the distortion
probability \eqn{\alpha_\ell}. The default proposal distribution is a
reflected random walk.
}
\concept{Metropolis-Hastings Update Functions}
