% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{alpha_prior_tests}
\alias{alpha_prior_tests}
\title{Tests for validity of \code{alpha_pior} argument.}
\usage{
alpha_prior_tests(alpha_prior, key_vars)
}
\arguments{
\item{alpha_prior}{A list of lists. A list of beta distribution shape
parameters representing the prior for the distortion
probabilities.}

\item{key_vars}{A character vector. The names of the fields in the record
data.}
}
\value{
Exception and error message if \code{alpha_prior} is invalid.
}
\description{
Both \code{\link{recursiveRL}} and \code{\link{regressionRL}} require an
\code{alpha_pior} argument that encodes the beta prior on the distortion
probabilites. \code{alpha_prior_tests} runs tests to make sure
\code{alpha_pior} is a valid data structure.
}
\section{Details}{

\code{alpha_pior} must be a list of lists. Each list in \code{alpha_pior}
must be named after an element of \code{key_vars}, and must contain two
strictly positive numeric elements \code{alpha} and \code{beta}.
\code{alpha_prior_tests} runs tests to make sure \code{alpha_prior} is
structured according to these standards.
}

\examples{
# creating uniform prior for RL500

key_vars <- c("fname_c1", "lname_c1", "by", "bm", "bd")

alpha_prior <- list(fname_c1 = list(alpha = 1, beta = 1),
                    lname_c1 = list(alpha = 1, beta = 1),
                    by       = list(alpha = 1, beta = 1),
                    bm       = list(alpha = 1, beta = 1),
                    bd       = list(alpha = 1, beta = 1))

alpha_prior_tests(key_vars, alpha_prior)

}
\seealso{
Other Prior Functions.: 
\code{\link{generate_alpha_prior}()}
}
\concept{Prior Functions.}
