% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{alpha_proposal_tests}
\alias{alpha_proposal_tests}
\title{Tests for validity of \code{alpha_proposal} argument.}
\usage{
alpha_proposal_tests(alpha_proposal, key_vars)
}
\arguments{
\item{alpha_proposal}{A list. A list containing the proposal distribution to
use in the Metropolis-Hastings step of
\code{\link{recursiveRL}} and \code{\link{regressionRL}}.}

\item{key_vars}{A character vector. The names of the fields in the record
data.}
}
\value{
Exception and error message if \code{alpha_proposal} is invalid.
}
\description{
\code{alpha_proposal_tests} performs tests to make sure \code{alpha_proposal}
is properly formatted for use in either \code{\link{recursiveRL}} or
\code{\link{regressionRL}}.
}
\section{Details}{

\code{alpha_proposal} must be either a list containing a string element
named \code{proposal}, which specifies the type of proposal distribution to
use for the distortion probabilities \eqn{\alpha}, and a non-negative numeric
element named \code{drift}, which controls the correlation of the proposals.
Otherwise, \code{alpha_proposal} can be a list of lists, with each sublist being
named after a value in \code{key_vars} and each containing separate \code{proposal}
and \code{drift} parameters. Combinations are also allowed, so that if some value
in \code{key_vars} does not appear in \code{alpha_proposal},
\code{alpha_proposal[["proposal"]]} and \code{alpha_proposal[["drift"]]} will be used
for the missing values.
}

\examples{
# creating reflected random walk proposal for RL500
key_vars <- c("fname_c1", "lname_c1", "by", "bm", "bd")

# one proposal for all fields
alpha_proposal <- list(proposal = "RRW", drift = 0.1)
alpha_proposal_tests(alpha_proposal, key_vars)

# different proposal for each field
alpha_proposal <- list(fname_c1 = list(proposal = "RRW", drift = 0.1),
                       lname_c1 = list(proposal = "RRW", drift = 0.2),
                       by       = list(proposal = "RRW", drift = 0.3),
                       bm       = list(proposal = "RRW", drift = 0.4),
                       bd       = list(proposal = "RRW", drift = 0.5))
alpha_proposal_tests(alpha_proposal, key_vars)

}
\seealso{
\code{\link{get_alpha_proposals}} for currently implemented values of
\code{proposal}.
}
