% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_tools.R
\name{beta_optim}
\alias{beta_optim}
\title{Optimization function for backsolving shape parameters for a beta prior.}
\usage{
beta_optim(par, mn, var)
}
\arguments{
\item{par}{Parameters to optimize over.}

\item{mn}{A float. The mean condition of the beta prior. Must be strictly
positive.}

\item{var}{A float. The variance condition of the beta prior. Must be
strictly positive.}
}
\value{
A vector of residuals.
}
\description{
Optimization function for backsolving shape parameters for a beta prior.
}
