% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_tools.R
\name{beta_solve}
\alias{beta_solve}
\title{Solve for the shape parameters of beta distribtution given mean and variance.}
\usage{
beta_solve(mn, var)
}
\arguments{
\item{mn}{A float. The mean condition of the beta prior. Must be strictly
positive.}

\item{var}{A float. The variance condition of the beta prior. Must be
strictly positive.}
}
\value{
A list. \eqn{(\alpha, \beta)}
}
\description{
\code{beta_solve} backsolves for \eqn{\alpha} and \eqn{\beta} of a beta
distribution given a mean and variance.
}
\examples{
beta_solve(mn = 2.6, var = 3.4)

}
