% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution.R
\name{emp_dist}
\alias{emp_dist}
\title{Calculates empirical frequencies of each value in a collection of records.}
\usage{
emp_dist(data, key_vars)
}
\arguments{
\item{data}{A data.frame. The data containing the records.}

\item{key_vars}{A character vector. The names of the columns from \code{data}
to use as key variables.}
}
\value{
A list of data.frames. Each list contains an index of values each key
  variable can take on and the corresponding empirical frequency.
}
\description{
\code{emp_dist} takes a data.frame of records \code{data} and \code{key_vars}
a vector of colmun names, then returns a list of data.frames (one for each
key variable), each containing a column of values and a column of
corresponding empirical frequencies.
}
