% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{generate_alpha_prior}
\alias{generate_alpha_prior}
\title{Generate prior for distortion probability \eqn{\alpha}.}
\usage{
generate_alpha_prior(key_vars)
}
\arguments{
\item{key_vars}{A character vector. The names of the fields in the record
data.}
}
\value{
A formatted list encoding the prior for \eqn{\alpha}.
}
\description{
Given a vector of key variable names, \code{generate_alpha_prior} prompts
the user to enter the parameters of the prior on the distortion
probability. The list that \code{generate_alpha_prior} can then directly
be used as the \code{alpha_prior} argument in either
\code{\link{recursiveRL}} or \code{\link{regressionRL}}.
}
\examples{
key_vars <- c("fname_c1", "lname_c1", "by", "bm", "bd")
alpha_prior <- generate_alpha_prior(key_vars)

}
\seealso{
Other Prior Functions.: 
\code{\link{alpha_prior_tests}()}
}
\concept{Prior Functions.}
