% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{get_alpha_proposals}
\alias{get_alpha_proposals}
\title{List the implemented proposal distributions for the distortion probability
\eqn{\alpha}.}
\usage{
get_alpha_proposals(verbose = TRUE)
}
\arguments{
\item{verbose}{A logical. If TRUE, the list of implemented proposal
distributions will be printed}
}
\value{
A character vector of the implemented proposal distributions.
}
\description{
Returns a list of the proposal distributions that are currently
implemented for generating new values of  the distortion probability
\eqn{\alpha} in \code{\link{alpha_metropolis}}.
}
