% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{hitmiss}
\alias{hitmiss}
\title{Likelihood for the \eqn{\ell}th field of the hit-and-miss model.}
\usage{
hitmiss(v, v_true, theta, alpha)
}
\arguments{
\item{v}{A string. The potential value for the \eqn{\ell^{th}} field of
the record that \code{hitmiss} calculates the likelihood for.}

\item{v_true}{A string. The true value of the \eqn{\ell^{th}} field of the
record.}

\item{theta}{A float. The empirical frequency of \code{v} in the set of
records.}

\item{alpha}{A float. The distortion probability for the \eqn{\ell^{th}}
field.}
}
\value{
The likelihood of \code{v}.
}
\description{
\code{hitmiss} gives the likelihood of the \eqn{\ell^{th}} field of a record
taking the value \code{v} given that the true value for the field is
\code{v_true}, the empirical frequency of \code{v} in all of the records is
\eqn{\theta_{v,\ell}}, and the distortion probability for the \eqn{\ell^{th}}
field is \eqn{\alpha_{\ell}}.
}
\section{Details}{

This function is used in the Metropolis-Hastings update step for the linkage
structure \eqn{\Lambda}.

Given that the true value for the \eqn{\ell^{th}} field is \code{v_true}, the
empirical frequency of \code{v} in all of the records is
\eqn{\theta_{v,\ell}}, and the distortion probability for the \eqn{\ell^{th}}
field is \eqn{\alpha_{\ell}}, then the likelihood of \code{v} is given by
\deqn{(1-\alpha_{\ell})\delta(v,v_{true}) + \alpha_{\ell} \theta_{v,\ell},}
where \eqn{\delta(v,v_{true}) = 1} if \eqn{v = v_{true}}, and
\eqn{\delta(v,v_{true}) = 0} otherwise.
}

\seealso{
Other Cluster Likelihood Functions: 
\code{\link{log_distortion_cluster}()},
\code{\link{log_joint_lm_cluster}()},
\code{\link{log_linkage_cluster}()},
\code{\link{log_lm_cluster}()},
\code{\link{recursive_cluster}()}
}
\concept{Cluster Likelihood Functions}
