% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{lambda_prior_tests}
\alias{lambda_prior_tests}
\title{Tests for validity of \code{lambda_pior} argument.}
\usage{
lambda_prior_tests(lambda_prior)
}
\arguments{
\item{lambda_prior}{A list. A list containing the type of prior to use for
the linkage structure and the corresponding parameters
for the prior.}
}
\value{
Exception and error message if \code{lambda_prior} is invalid.
}
\description{
Both \code{\link{recursiveRL}} and \code{\link{regressionRL}} require an
\code{lambda_pior} argument that encodes the prior on the linkage
structure \eqn{\Lambda}. \code{lambda_prior_tests} runs tests to make sure
\code{lambda_pior} is a valid data structure.
}
\section{Details}{

\code{lambda_prior} must include an element named "prior" that contains a
string specifying the prior to be used for the linkage structure. Currently
the values 'uniform' and 'PYP' are supported. If 'prior' = 'PYP', then
\code{lambda_prior} must also contain two numeric arguments 'nu' and
'sigma'. See \code{\link{pyp_mean}} and \code{\link{pyp_var}} for more
detail.
}

\examples{

uni_prior <- list(prior = "uniform")
pyp_params <- pyp_solve(mn = 450, var = 500, N = 500)
pyp_prior <- list(prior = "PYP",
                  nu = pyp_params$nu,
                  sigma = pyp_params$sigma)

lapply(list(uni_prior, pyp_prior), lambda_prior_tests)

}
