% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{lgratio}
\alias{lgratio}
\title{Calculate \eqn{\log \frac{\Gamma(a + b)}{\Gamma(a)}}.}
\usage{
lgratio(a, b)
}
\arguments{
\item{a, b}{A float. \eqn{a > 0} and \eqn{b > -a}.}
}
\value{
\eqn{\log \frac{\Gamma(a + b)}{\Gamma(a)}}
}
\description{
Calculate \eqn{\log \frac{\Gamma(a + b)}{\Gamma(a)}}.
}
\seealso{
Other PYP Functions: 
\code{\link{pyp_mean}()},
\code{\link{pyp_optim}()},
\code{\link{pyp_solve}()},
\code{\link{pyp_var}()},
\code{\link{pyp_warnings}()}
}
\concept{PYP Functions}
