% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_beta_prior_dens}
\alias{log_beta_prior_dens}
\title{Log density of \code{beta} given a prior.}
\usage{
log_beta_prior_dens(beta, beta_prior)
}
\arguments{
\item{beta}{A numeric vector. The sample of \eqn{\beta} to
calculate the log density for}

\item{beta_prior}{A list. A list encoding the prior for \eqn{\beta}.}
}
\value{
Log density of \code{beta}.
}
\description{
Log density of \code{beta} given a prior.
}
