% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{log_joint_lm_cluster}
\alias{log_joint_lm_cluster}
\title{Joint conditional regression log-likelihood.}
\usage{
log_joint_lm_cluster(linkage, model_data, beta, sigma_data, sigma_y, sigma_x)
}
\arguments{
\item{linkage}{An integer vector. The cluster labels for each record.}

\item{model_data}{A data.frame. The model.frame of the regression. The first
column must contain the outcome variable \eqn{y} and the
remaining columns must contain the model covariates \eqn{X}.}

\item{beta}{A numeric vector. The regression coefficients. Can also be
a single float in the case of univariate regression.}

\item{sigma_data}{A matrix. The variance-covariance matrix of the covariates.
Must be symetric positive-definite. Can also be a strictly
positive float in the case of univariate regression.}

\item{sigma_y}{A float. The variance of the outcome variable \eqn{y}. Must
be strictly positive.}

\item{sigma_x}{A matrix. The current variance-covariance matrix of the
distortion of the covariates. Must be symetric
positive-definite. Can also be a strictly positive float
in the case of univariate regression.}
}
\value{
The joint log-likelihood of \eqn{(y,X)}.
}
\description{

}
\section{Details}{

Let \eqn{C_j} be the cluster of records corresponding to the \eqn{j^{th}} true
entity, then assuming that there are \eqn{N_{pop}} true entities represented
in the records (i.e. there are \eqn{N_{pop}} unique elements of
\code{linkage}), \code{log_joint_lm_cluster} calculates
\deqn{\sum_{j=1}^{N_{pop}} \log P\left([y,X]_{C_j} \; | \; \Lambda, \beta,
\Sigma_{\widetilde{x}}, \Sigma_{x|\widetilde{x}},
\sigma^2_{y|\widetilde{x}}\right).}
}

\seealso{
Other Cluster Likelihood Functions: 
\code{\link{hitmiss}()},
\code{\link{log_distortion_cluster}()},
\code{\link{log_linkage_cluster}()},
\code{\link{log_lm_cluster}()},
\code{\link{recursive_cluster}()}
}
\concept{Cluster Likelihood Functions}
