% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{log_linkage_cluster}
\alias{log_linkage_cluster}
\title{Cluster log-likelihood.}
\usage{
log_linkage_cluster(cluster, thetas, alphas, key_vars)
}
\arguments{
\item{cluster}{A data.frame. All of the records and their values for each key
variable for a specific cluster.}

\item{thetas}{A list of matrices. Each matrix corresponds to a key variable
and contains two columns: 1) all of the values each key variable
can take in the data, and 2) the corresponding empirical
frequency of each value.}

\item{alpha}{A data.frame. Contains a column \code{key_vars} with each of the
key variables, and a column \code{alpha} with the corresponding
distortion probabilities.}
}
\value{
The joint log-likelihood of the provided cluster.
}
\description{
Given a data.frame of records for a specific cluster,
\code{log_recursive_cluster} loops over each key variable (each column of the
cluster), calls \code{recursive_cluster}, and returns the sums of the
log-likelihoods for each field.
}
\section{Details}{

This function is used in the Metropolis-Hastings update step for the linkage
structure \eqn{\Lambda}.
}

\seealso{
\code{\link{recursive_cluster}} for the actual recursion formula
  implemented for this function.

Other Cluster Likelihood Functions: 
\code{\link{hitmiss}()},
\code{\link{log_distortion_cluster}()},
\code{\link{log_joint_lm_cluster}()},
\code{\link{log_lm_cluster}()},
\code{\link{recursive_cluster}()}
}
\concept{Cluster Likelihood Functions}
