% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{log_lm_cluster}
\alias{log_lm_cluster}
\title{Conditional regression log-likelihood for single cluster.}
\usage{
log_lm_cluster(cluster, B, Sigma)
}
\arguments{
\item{cluster}{A data.fram. The model.frame of the regression corresponding
to the records of a particular cluster. The first column must
contain the outcome variable \eqn{y} and the remaining columns
must contain the model covariates \eqn{X}.}

\item{B}{A matrix. The first block component of the variance-covariance
matrix.}

\item{Sigma}{A matrix. The first block component of the variance-covariance
matrix.}
}
\value{
The conditional log-likelihood of the given cluster.
}
\description{

}
\section{Details}{

Let \eqn{C_j} be the cluster of records corresponding to the \eqn{j^{th}}
true entity, then assuming that there are \eqn{n} records in \eqn{C_j}, and
\eqn{p} covariates in \eqn{X}, \code{log_lm_cluster} calculates
\deqn{\log P\left([y,X]_{C_j} \; | \; \beta, \Sigma_{\widetilde{x}},
\Sigma_{x|\widetilde{x}}, \sigma^2_{y|\widetilde{x}}\right) =
\log N_{n(p+1)}\left(\mathbf{0}, \; (\mathbf{1}_n \mathbf{1}_n') \otimes
\mathbf{B} + \mathbf{I}_{n} \otimes \Sigma \right).}
where
\eqn{\mathbf{B}=}
\tabular{rccl}{
| \tab \eqn{\beta' \Sigma_{\widetilde{x}} \beta} \tab \eqn{\beta'
\Sigma_{\widetilde{x}}} \tab | \cr
| \tab \eqn{\Sigma_{\widetilde{x}} \beta} \tab \eqn{\Sigma_{\widetilde{x}}}
\tab |
}
and
\eqn{\Sigma=}
\tabular{lccr}{
| \tab \eqn{\sigma^2_{y|\widetilde{x}}} \tab \eqn{\mathbf{0}} \tab | \cr
| \tab \eqn{\mathbf{0}} \tab \eqn{\Sigma_{x|\widetilde{x}}} \tab |.
}
}

\seealso{
Other Cluster Likelihood Functions: 
\code{\link{hitmiss}()},
\code{\link{log_distortion_cluster}()},
\code{\link{log_joint_lm_cluster}()},
\code{\link{log_linkage_cluster}()},
\code{\link{recursive_cluster}()}
}
\concept{Cluster Likelihood Functions}
