% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{log_p_lambda}
\alias{log_p_lambda}
\title{Log prior probability of linkage struture}
\usage{
log_p_lambda(lambda, record, linkage, lambda_prior)
}
\arguments{
\item{lambda}{An integer. The cluster label corresponding to the record
in the \code{record} row of the data.}

\item{record}{An integer. The row corresponding to the current record
being updated in the data.}

\item{linkage}{An integer vector. The current cluster labels for each
record.}

\item{lambda_prior}{A list. A list containing the type of prior to use for
the linkage structure and the corresponding parameters
for the prior.}
}
\value{
\eqn{\log P(\lambda_{ij} \;|\; \lambda_{-(ij)})}
}
\description{
\code{log_p_lambda} calculates \eqn{\log P(\lambda_{ij} \;|\;
\lambda_{-(ij)})} where \eqn{\lambda_{ij}} is the cluster label of the
\eqn{j^{th}} record in the \eqn{i^{th}} database.
}
