% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_sigma_data_prior_dens}
\alias{log_sigma_data_prior_dens}
\title{Log density of \code{sigma_data} given a prior.}
\usage{
log_sigma_data_prior_dens(sigma_data, sigma_data_prior)
}
\value{
Log density of \code{sigma_data}.
}
\description{
Log density of \code{sigma_data} given a prior.
}
