% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_sigma_x_prior_dens}
\alias{log_sigma_x_prior_dens}
\title{Log density of \code{sigma_x} given a prior.}
\usage{
log_sigma_x_prior_dens(sigma_x, sigma_x_prior)
}
\value{
Log density of \code{sigma_x}.
}
\description{
Log density of \code{sigma_x} given a prior.
}
