% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_sigma_y_prior_dens}
\alias{log_sigma_y_prior_dens}
\title{Log density of \code{sigma_y} given a prior.}
\usage{
log_sigma_y_prior_dens(sigma_y, sigma_y_prior)
}
\arguments{
\item{sigma_y}{A numeric vector. The sample of \eqn{\sigma^_y} to
calculate the log density for.}

\item{sigma_y_prior}{A list. A list encoding the prior for \eqn{\sigma^_y}.}
}
\value{
Log density of \code{sigma_y}.
}
\description{
Log density of \code{sigma_y} given a prior.
}
