% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{pyp_mean}
\alias{pyp_mean}
\title{Calculate the mean of a Pittman-Yor process.}
\usage{
pyp_mean(nu, sigma, N)
}
\arguments{
\item{nu, sigma}{A float. PYP parameters such that \eqn{\sigma \in [0,1)} and
\eqn{\nu > -\sigma} or \eqn{\sigma < 0} and \eqn{\nu > m|\sigma|}
for some integer \eqn{m}.}

\item{N}{An integer. The number of records.}
}
\value{
Mean of PYP prior.
}
\description{
Given \eqn{N} entities (in this case records), \code{pyp_mean} calculates the
expected number of clusters for a Pittman-Yor process with parameters
\eqn{\sigma} and \eqn{\nu}, \deqn{\frac{\nu}{\sigma} \left[\frac{(\nu +
\sigma)_{N\uparrow}}{\nu_{N\uparrow}} - 1\right]} where \eqn{x_{s\uparrow} =
\frac{\Gamma(x+s)}{\Gamma(x)}}.
}
\seealso{
Other PYP Functions: 
\code{\link{lgratio}()},
\code{\link{pyp_optim}()},
\code{\link{pyp_solve}()},
\code{\link{pyp_var}()},
\code{\link{pyp_warnings}()}
}
\concept{PYP Functions}
