% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{pyp_optim}
\alias{pyp_optim}
\title{Optimization function for backsolving parameters of a PYP prior.}
\usage{
pyp_optim(par, mn, var, N)
}
\arguments{
\item{par}{Parameters to optimize over.}

\item{mn}{An integer. The mean condition of the PYP prior.}

\item{var}{A float. The variance condition of the PYP prior. Must be strictly
positive.}

\item{N}{An integer. The number of records.}
}
\value{
A vector of residuals.
}
\description{
Gives non-linear equations for \eqn{\nu} and \eqn{\sigma} in terms of a known
mean, variance, and \eqn{N}.
}
\seealso{
Other PYP Functions: 
\code{\link{lgratio}()},
\code{\link{pyp_mean}()},
\code{\link{pyp_solve}()},
\code{\link{pyp_var}()},
\code{\link{pyp_warnings}()}
}
\concept{PYP Functions}
