% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{pyp_solve}
\alias{pyp_solve}
\title{Solve for parameters a PYP prior given a mean, variance, and \eqn{N}.}
\usage{
pyp_solve(mn, var, N)
}
\arguments{
\item{mn}{An integer. The mean condition of the PYP prior.}

\item{var}{A float. The variance condition of the PYP prior. Must be
strictly positive.}

\item{N}{An integer. The number of records.}
}
\value{
A list. \eqn{(\sigma, \nu).}
}
\description{
\code{pyp_solve} finds the parameters \eqn{\nu} and \eqn{\sigma} that give a
PYP prior given the mean \code{mn}, variance \code{var}, for a given number
of entities \eqn{N}.
}
\examples{
pyp_solve(mn = 450, var = 500, N = 500)
}
\seealso{
Other PYP Functions: 
\code{\link{lgratio}()},
\code{\link{pyp_mean}()},
\code{\link{pyp_optim}()},
\code{\link{pyp_var}()},
\code{\link{pyp_warnings}()}
}
\concept{PYP Functions}
