% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{pyp_var}
\alias{pyp_var}
\title{Calculate the variance of a Pittman-Yor process.}
\usage{
pyp_var(nu, sigma, N)
}
\arguments{
\item{nu, sigma}{A float. The PYP parameters such that \eqn{\sigma \in [0,1)}
and \eqn{\nu > - \sigma} or \eqn{\sigma < 0} and
\eqn{\nu > m|\sigma|} for some integer \eqn{m}.}

\item{N}{An integer. The number of records.}
}
\value{
Variance of PYP prior.
}
\description{
Given \eqn{N} entities (in this case records), \code{pyp_var} calculates the
variance of the number of clusters for a Pittman-Yor process with parameters
\eqn{\sigma} and \eqn{\nu},
\deqn{\frac{\nu(\nu+\sigma)}{\sigma^2}\frac{(\nu+2\sigma)_{N\uparrow}}{\nu_N\uparrow}
- \left[\frac{\nu}{\sigma}\frac{(\nu +
\sigma)_{N\uparrow}}{\nu_{N\uparrow}}\right]^2 - \frac{\nu}{\sigma}\frac{(\nu
+ \sigma)_{N\uparrow}}{\nu_{N\uparrow}}} where \eqn{x_{s\uparrow} =
\frac{\Gamma(x+s)}{\Gamma(x)}}.
}
\seealso{
Other PYP Functions: 
\code{\link{lgratio}()},
\code{\link{pyp_mean}()},
\code{\link{pyp_optim}()},
\code{\link{pyp_solve}()},
\code{\link{pyp_warnings}()}
}
\concept{PYP Functions}
