% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyp_tools.R
\name{pyp_warnings}
\alias{pyp_warnings}
\title{PYP function parameter tests.}
\usage{
pyp_warnings(nu, sigma, N)
}
\arguments{
\item{nu, sigma}{A float. PYP parameters such that \eqn{\sigma \in [0,1)} and
\eqn{\nu > -\sigma} or \eqn{\sigma < 0} and \eqn{\nu > m|\sigma|}
for some integer \eqn{m}.}

\item{N}{An integer. The number of records.}
}
\value{
Warning if broken parameters are provided.
}
\description{
PYP function parameter tests.
}
\seealso{
Other PYP Functions: 
\code{\link{lgratio}()},
\code{\link{pyp_mean}()},
\code{\link{pyp_optim}()},
\code{\link{pyp_solve}()},
\code{\link{pyp_var}()}
}
\concept{PYP Functions}
