% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{recursive_cluster}
\alias{recursive_cluster}
\title{Cluster likelihood for the \eqn{\ell}th field.}
\usage{
recursive_cluster(cluster, thetas_l, alpha_l)
}
\arguments{
\item{cluster}{A vector. The vector of strings/factors for the
\eqn{\ell^{th}} field of the selected cluster.}

\item{thetas_l}{A matrix. The first column of \code{thetas_l} must be the
possible values that \code{cluster}, and the second column
should contain the corresponding frequencies of each value
in the whole set of records.}

\item{alpha_l}{A float. The distortion probability of the \eqn{\ell^{th}}
field.}
}
\value{
The joint likelihood of the \eqn{\ell^{th}} field of a cluster.
}
\description{
Given the \eqn{\ell^{th}} field of a cluster as a vector along with the
empirical frequencies \eqn{\theta_\ell} and distortion probability,
\code{recursive_cluster} uses a recursive formula to calculate the joint
likelihood for the \eqn{q^{th}} cluster \eqn{v_{C_q}}.
}
\section{Details}{

This function is used in the Metropolis-Hastings update step for the linkage
structure \eqn{\Lambda}.

The recursive formula for \eqn{P(v_{C_{q},\ell} \;|\; \alpha_\ell, \lambda)}
is given by \deqn{P(v_{C_{q},\ell} \;|\; \alpha_\ell, \lambda) =
\theta_{v_{ij\ell}},} if \eqn{v_{ij\ell} = v_{C_q,\ell},} and

\eqn{P(v_{C_{q},\ell} \;|\; \alpha_\ell, \lambda) =}
\deqn{(1-\alpha_\ell)\left[\prod_{(i', j') \in C_{q\setminus(ij)}} (1 -
\alpha_\ell) \delta(v_{i' j'\ell}, v_{ij\ell}) + \alpha_\ell \theta_{v_{i'
j'},\ell}\right] \theta_{v_{ij\ell}} + \alpha_\ell \theta_{v_{ij\ell}}
P(V_{C_{q}\setminus(ij),\ell} \; | \; \alpha_\ell, \lambda)} otherwise.
}

\seealso{
Other Cluster Likelihood Functions: 
\code{\link{hitmiss}()},
\code{\link{log_distortion_cluster}()},
\code{\link{log_joint_lm_cluster}()},
\code{\link{log_linkage_cluster}()},
\code{\link{log_lm_cluster}()}
}
\concept{Cluster Likelihood Functions}
