% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveRL.R
\name{recursiveRL}
\alias{recursiveRL}
\title{Title}
\usage{
recursiveRL(
  data,
  key_vars,
  sample_size,
  burnin = 0,
  thin = 1,
  n_chains = 1,
  progressbar = TRUE,
  alpha_init = NULL,
  alpha_prior,
  alpha_proposal = NULL,
  lambda_init = NULL,
  lambda_prior
)
}
\arguments{
\item{lambda_prior}{}
}
\value{

}
\description{
Title
}
