% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresionRL.R
\name{regressionRL}
\alias{regressionRL}
\title{Title}
\usage{
regressionRL(
  formula,
  data,
  key_vars,
  sample_size,
  burnin = 0,
  thin = 1,
  n_chains = 1,
  progressbar = TRUE,
  alpha_init = NULL,
  alpha_prior,
  alpha_proposal = NULL,
  lambda_init = NULL,
  lambda_prior,
  beta_init = NULL,
  beta_prior = NULL,
  beta_proposal = NULL,
  sigma_y_init = NULL,
  sigma_y_prior = NULL,
  sigma_y_proposal = NULL,
  sigma_x_diag = FALSE,
  sigma_x_init = NULL,
  sigma_x_prior = NULL,
  sigma_x_proposal = NULL,
  sigma_data_fixed = TRUE,
  sigma_data_diag = FALSE,
  sigma_data_init = NULL,
  sigma_data_prior = NULL,
  sigma_data_proposal = NULL
)
}
\arguments{
\item{sigma_data_proposal}{}
}
\value{

}
\description{
Title
}
