% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_functions.R
\name{sigma_x_metropolis}
\alias{sigma_x_metropolis}
\title{Metropolis-Hastings update of \eqn{\Sigma_{x|\widetilde{x}}}.}
\usage{
sigma_x_metropolis(
  linkage,
  model_data,
  beta,
  sigma_data,
  sigma_y,
  sigma_x,
  sigma_x_prior,
  sigma_x_proposal
)
}
\arguments{
\item{linkage}{An integer vector. The cluster labels for each record.}

\item{model_data}{A data.frame. The model.frame of the regression. The
first column must contain the outcome variable \eqn{y}
and the remaining columns must contain the model
covariates \eqn{X}.}

\item{beta}{A numeric vector. The value of the regression
coefficients. Can also be a single float in the case of
univariate regression.}

\item{sigma_data}{A matrix. The variance-covariance matrix of the
covariates. Must be symetric positive-definite. Can also
be a strictly positive float in the case of univariate
regression.}

\item{sigma_y}{A float. The variance of the outcome variable
\eqn{y}. Must be strictly positive.}

\item{sigma_x}{A matrix. The current variance-covariance matrix of the
distortion of the covariates. Must be symetric
positive-definite. Can also be a strictly positive float
in the case of univariate regression.}

\item{sigma_x_prior}{A list. A list encoding the prior distribution of
\eqn{\Sigma_{x|\widetilde{x}}}.}

\item{sigma_x_proposal}{A list. A list encoding the proposal distribution to use
for \eqn{\Sigma_{x|\widetilde{x}}} in the
Metropolis-Hastings step.}
}
\value{
An updated value of \eqn{\Sigma_{x|\widetilde{x}}}.
}
\description{
Implementation of the Metropolis-Hastings update step for the covariate
distortion variance \eqn{\Sigma_{x|\widetilde{x}}}.
}
